pyflask/day09/cmdscript01/requirements.txt
flask
flask-script
flask-sqlalchemy
pymysql
flask-migrate
